requirements.txt
imgurpython
pillow
